Refactoring client class
