+++ AdobeStockClient/Model/Client.php
@@ -10,6 +10,7 @@ namespace Magento\AdobeStockClient\Model;
 
 use AdobeStock\Api\Client\AdobeStock;
 use AdobeStock\Api\Core\Constants;
+use AdobeStock\Api\Exception\StockApi;
 use AdobeStock\Api\Models\SearchParameters;
 use AdobeStock\Api\Models\StockFile;
 use AdobeStock\Api\Request\SearchFiles as SearchFilesRequest;
@@ -17,14 +18,11 @@ use AdobeStock\Api\Response\License;
 use Exception;
 use Magento\AdobeImsApi\Api\Data\ConfigInterface as ImsConfig;
 use Magento\AdobeImsApi\Api\UserProfileRepositoryInterface;
+use Magento\AdobeStockClient\Model\StockFileToDocument;
 use Magento\AdobeStockClientApi\Api\ClientInterface;
 use Magento\AdobeStockClientApi\Api\Data\ConfigInterface;
 use Magento\AdobeStockClientApi\Api\SearchParameterProviderInterface;
 use Magento\Authorization\Model\UserContextInterface;
-use Magento\Framework\Api\AttributeValue;
-use Magento\Framework\Api\AttributeValueFactory;
-use Magento\Framework\Api\Search\DocumentFactory;
-use Magento\Framework\Api\Search\DocumentInterface;
 use Magento\Framework\Api\Search\SearchResultFactory;
 use Magento\Framework\Api\Search\SearchResultInterface;
 use Magento\Framework\Api\Search\SearchCriteriaInterface;
@@ -58,14 +56,9 @@ class Client implements ClientInterface
     private $searchResultFactory;
 
     /**
-     * @var DocumentFactory
+     * @var StockFileToDocument
      */
-    private $documentFactory;
-
-    /**
-     * @var AttributeValueFactory
-     */
-    private $attributeValueFactory;
+    private $stockFileToDocument;
 
     /**
      * @var SearchParameterProviderInterface
@@ -106,9 +99,7 @@ class Client implements ClientInterface
      * Client constructor.
      * @param ConfigInterface $clientConfig
      * @param ImsConfig $imsConfig
-     * @param DocumentFactory $documentFactory
      * @param SearchResultFactory $searchResultFactory
-     * @param AttributeValueFactory $attributeValueFactory
      * @param SearchParameterProviderInterface $searchParametersProvider
      * @param LocaleResolver $localeResolver
      * @param ConnectionFactory $connectionFactory
@@ -116,26 +107,24 @@ class Client implements ClientInterface
      * @param LoggerInterface $logger
      * @param UserProfileRepositoryInterface $userProfileRepository
      * @param UserContextInterface $userContext
+     * @param StockFileToDocument $stockFileToDocument
      */
     public function __construct(
         ConfigInterface $clientConfig,
         ImsConfig $imsConfig,
-        DocumentFactory $documentFactory,
         SearchResultFactory $searchResultFactory,
-        AttributeValueFactory $attributeValueFactory,
         SearchParameterProviderInterface $searchParametersProvider,
         LocaleResolver $localeResolver,
         ConnectionFactory $connectionFactory,
         LicenseRequestFactory $licenseRequestFactory,
         LoggerInterface $logger,
         UserProfileRepositoryInterface $userProfileRepository,
-        UserContextInterface $userContext
+        UserContextInterface $userContext,
+        StockFileToDocument $stockFileToDocument
     ) {
         $this->clientConfig = $clientConfig;
         $this->imsConfig = $imsConfig;
-        $this->documentFactory = $documentFactory;
         $this->searchResultFactory = $searchResultFactory;
-        $this->attributeValueFactory = $attributeValueFactory;
         $this->searchParametersProvider = $searchParametersProvider;
         $this->localeResolver = $localeResolver;
         $this->connectionFactory = $connectionFactory;
@@ -143,6 +132,7 @@ class Client implements ClientInterface
         $this->logger = $logger;
         $this->userProfileRepository = $userProfileRepository;
         $this->userContext = $userContext;
+        $this->stockFileToDocument = $stockFileToDocument;
     }
 
     /**
@@ -162,7 +152,7 @@ class Client implements ClientInterface
             $response = $connection->getNextResponse();
             /** @var StockFile $file */
             foreach ($response->getFiles() as $file) {
-                $items[] = $this->convertStockFileToDocument($file);
+                $items[] = $this->stockFileToDocument->convert($file);
             }
             $totalCount = $response->getNbResults();
         } catch (Exception $exception) {
@@ -185,6 +175,8 @@ class Client implements ClientInterface
      *
      * @param int $contentId
      * @return License
+     * @throws IntegrationException
+     * @throws StockApi
      */
     private function getLicenseInfo(int $contentId): License
     {
@@ -213,39 +205,11 @@ class Client implements ClientInterface
     }
 
     /**
-     * Convert a stock file object to a document object
-     *
-     * @param StockFile $file
-     * @return DocumentInterface
-     * @throws IntegrationException
-     */
-    private function convertStockFileToDocument(StockFile $file): DocumentInterface
-    {
-        $itemData = (array) $file;
-        $itemId = $itemData['id'];
-
-        $category = (array) $itemData['category'];
-
-        $itemData['category'] = $category;
-        $itemData['category_id'] = $category['id'];
-        $itemData['category_name'] = $category['name'];
-
-        $attributes = $this->createAttributes('id', $itemData);
-
-        $item = $this->documentFactory->create();
-        $item->setId($itemId);
-        $item->setCustomAttributes($attributes);
-
-        return $item;
-    }
-
-    /**
      * Create and return search request based on search criteria
      *
      * @param SearchCriteriaInterface $searchCriteria
      * @return SearchFilesRequest
-     * @throws IntegrationException
-     * @throws \AdobeStock\Api\Exception\StockApi
+     * @throws StockApi
      */
     private function getSearchRequest(SearchCriteriaInterface $searchCriteria): SearchFilesRequest
     {
@@ -277,47 +241,6 @@ class Client implements ClientInterface
         }
 
         return $resultColumnArray;
-    }
-
-    /**
-     * Create custom attributes for columns returned by search
-     *
-     * @param string $idFieldName
-     * @param array $itemData
-     * @return AttributeValue[]
-     * @throws IntegrationException
-     */
-    private function createAttributes(string $idFieldName, array $itemData): array
-    {
-        try {
-            $attributes = [];
-
-            $idFieldNameAttribute = $this->attributeValueFactory->create();
-            $idFieldNameAttribute->setAttributeCode('id_field_name');
-            $idFieldNameAttribute->setValue($idFieldName);
-            $attributes['id_field_name'] = $idFieldNameAttribute;
-
-            foreach ($itemData as $key => $value) {
-                if ($value === null) {
-                    continue;
-                }
-                $attribute = $this->attributeValueFactory->create();
-                $attribute->setAttributeCode($key);
-                if (is_bool($value)) {
-                    // for proper work of form and grid (for example for Yes/No properties)
-                    $value = (string)(int)$value;
-                }
-                $attribute->setValue($value);
-                $attributes[$key] = $attribute;
-            }
-            return $attributes;
-        } catch (Exception $exception) {
-            $message = __(
-                'Create attributes process failed: %error_message',
-                ['error_message' => $exception->getMessage()]
-            );
-            $this->processException($message, $exception);
-        }
     }
 
     /**
